ContentLine: Prevent plain_delivery_length truncation
plain_delivery_length is int64_t and for Content-Length > INT_MAX,
would be truncated and result in a negative value, crashing later
in DoDeliver(). Do the min with the wider type to prevent this.
Cast to int is safe as len is a positive int here.

Added testcase would segfault on current master.

Fixup for a8069fcfa75582da15f11f72813c280f2efcd89e.

Reported by @JordanBarnartt, test case created and fix suggested by Claude.

+++ src/analyzer/protocol/tcp/ContentLine.cc
@@ -154,7 +154,7 @@ void ContentLine_Analyzer::DoDeliver(int len, const u_char* data) {
         }
 
         if ( plain_delivery_length > 0 ) {
-            auto deliver_plain = std::min<int>(plain_delivery_length, len);
+            auto deliver_plain = static_cast<int>(std::min<int64_t>(plain_delivery_length, len));
 
             last_char = 0; // clear last_char
             plain_delivery_length -= deliver_plain;
